% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsclient.R
\name{auth_user_token}
\alias{auth_user_token}
\title{auth_user_token}
\usage{
auth_user_token(username, token, endpoint_url)
}
\arguments{
\item{username}{username}

\item{token}{token to authenticate}

\item{endpoint_url}{Endpoint to check}
}
\value{
TRUE if the server token and user is valid, FALSE otherwise
}
\description{
Check if the user token is valid
}
