% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstore.R
\name{check_return_type}
\alias{check_return_type}
\title{check_return_type}
\usage{
check_return_type(return_type)
}
\arguments{
\item{return_type}{The return type to check}
}
\description{
Stop if return_type is not valid
}
