% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsclient.R
\name{create_fastq_dataset_rs}
\alias{create_fastq_dataset_rs}
\title{create_fastq_dataset_rs}
\usage{
create_fastq_dataset_rs(
  client,
  name,
  description,
  qc_passed,
  paired_end,
  index_read,
  project_ids,
  project_names,
  metadata,
  fq_file_r1_id = NULL,
  fq_file_r2_id = NULL,
  fq_file_i1_id = NULL,
  fq_file_i2_id = NULL
)
}
\arguments{
\item{client}{ReadStore client}

\item{name}{Dataset name}

\item{description}{Dataset description}

\item{qc_passed}{QC Pass}

\item{paired_end}{Paired End}

\item{index_read}{Index Read}

\item{project_ids}{List of project IDs}

\item{project_names}{List of project names}

\item{metadata}{Metadata (must be a named list in R)}

\item{fq_file_r1_id}{Fastq file R1 ID}

\item{fq_file_r2_id}{Fastq file R2 ID}

\item{fq_file_i1_id}{Fastq file I1 ID}

\item{fq_file_i2_id}{Fastq file I2 ID}
}
\value{
A list containing the created Fastq dataset
}
\description{
Create fastq dataset from the ReadStore API
ID or Name must be provided
}
