% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsclient.R
\name{create_fq_file_rs}
\alias{create_fq_file_rs}
\title{create_fq_file_rs}
\usage{
create_fq_file_rs(
  client,
  name,
  read_type,
  qc_passed,
  read_length,
  num_reads,
  size_mb,
  qc_phred_mean,
  qc_phred,
  upload_path,
  md5_checksum,
  staging,
  pipeline_version
)
}
\arguments{
\item{client}{ReadStore client}

\item{name}{Fastq file name}

\item{read_type}{Read type (R1, R2, I1, I2)}

\item{qc_passed}{QC Pass}

\item{read_length}{Read length}

\item{num_reads}{Number of reads}

\item{size_mb}{Size in MB}

\item{qc_phred_mean}{QC Phred Mean}

\item{qc_phred}{QC Phred (must be a named list in R)}

\item{upload_path}{Upload Path}

\item{md5_checksum}{MD5 Checksum}

\item{staging}{Staging}

\item{pipeline_version}{Pipeline Version}
}
\value{
(json) list of projects (list objects)
}
\description{
Create fastq files from the ReadStore API
}
