% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstore.R
\name{create_project}
\alias{create_project}
\title{create_project}
\usage{
create_project(
  client,
  name,
  description = "",
  metadata = list(),
  dataset_metadata_keys = c()
)
}
\arguments{
\item{client}{ReadStore client}

\item{name}{Project name}

\item{description}{Project description}

\item{metadata}{List of project metadata}

\item{dataset_metadata_keys}{vector of dataset metadata keys}
}
\value{
The created project object
}
\description{
Create a new project by calling create_project_rs
}
