% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsclient.R
\name{create_project_rs}
\alias{create_project_rs}
\title{create_project_rs}
\usage{
create_project_rs(client, name, description, metadata, dataset_metadata_keys)
}
\arguments{
\item{client}{ReadStore client}

\item{name}{Project ID to return}

\item{description}{Project name to return}

\item{metadata}{List of project metadata}

\item{dataset_metadata_keys}{vector of dataset metadata keys}
}
\value{
json object (list) with project
}
\description{
Create project from the ReadStore API
}
