% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstore.R
\name{delete_dataset}
\alias{delete_dataset}
\title{delete_dataset}
\usage{
delete_dataset(client, dataset_id = NULL, dataset_name = NULL)
}
\arguments{
\item{client}{ReadStore client}

\item{dataset_id}{Dataset ID}

\item{dataset_name}{Dataset name}
}
\value{
ID of the deleted dataset
}
\description{
Delete a dataset by checking its existence and calling delete_fastq_dataset_rs
}
