% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsclient.R
\name{delete_fastq_dataset_rs}
\alias{delete_fastq_dataset_rs}
\title{delete_fq_file_rs}
\usage{
delete_fastq_dataset_rs(client, dataset_id)
}
\arguments{
\item{client}{ReadStore client}

\item{dataset_id}{Fastq file ID}
}
\value{
(json) list of projects (list objects)
}
\description{
Delete fastq file from the ReadStore API
}
