% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsclient.R
\name{delete_fq_file_rs}
\alias{delete_fq_file_rs}
\title{delete_fq_file_rs}
\usage{
delete_fq_file_rs(client, fq_file_id)
}
\arguments{
\item{client}{ReadStore client}

\item{fq_file_id}{Fastq file ID}
}
\value{
(json) list of projects (list objects)
}
\description{
Delete fastq file from the ReadStore API
}
