% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstore.R
\name{delete_pro_data}
\alias{delete_pro_data}
\title{delete_pro_data}
\usage{
delete_pro_data(
  client,
  pro_data_id = NULL,
  name = NULL,
  dataset_id = NULL,
  dataset_name = NULL,
  version = NULL
)
}
\arguments{
\item{client}{ReadStore client}

\item{pro_data_id}{ProData ID to return}

\item{name}{Name of ProData entry to filter}

\item{dataset_id}{Dataset ID to filter}

\item{dataset_name}{Dataset name to filter}

\item{version}{Version of ProData entry to filter}
}
\description{
Delete ProData entry from ReadStore
Delete by ProData Id or combination of Name and Dataset ID or Name
If version is NULL, delete the latest valid version
}
