% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstore.R
\name{delete_project}
\alias{delete_project}
\title{delete_project}
\usage{
delete_project(client, project_id = NULL, project_name = NULL)
}
\arguments{
\item{client}{ReadStore client}

\item{project_id}{Project ID to delete}

\item{project_name}{Project name to delete}
}
\value{
ID of the deleted project
}
\description{
Delete an existing project by checking its existence and calling delete_project_rs
}
