% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsclient.R
\name{delete_project_rs}
\alias{delete_project_rs}
\title{delete_project_rs}
\usage{
delete_project_rs(client, project_id)
}
\arguments{
\item{client}{ReadStore client}

\item{project_id}{Project ID}
}
\value{
(json) list of projects (list objects)
}
\description{
Delete Project from the ReadStore API
}
