% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstore.R
\name{download_dataset_attachment}
\alias{download_dataset_attachment}
\title{download_dataset_attachment}
\usage{
download_dataset_attachment(
  client,
  attachment_name,
  dataset_id = NULL,
  dataset_name = NULL,
  outpath = NULL
)
}
\arguments{
\item{client}{ReadStore client}

\item{attachment_name}{Attachment name to download}

\item{dataset_id}{Dataset ID to return}

\item{dataset_name}{Dataset name to return}

\item{outpath}{Path to save the attachment}
}
\description{
Download attachment from a dataset
ID or Name of dataset must be provided
Also provide the attachment name to download
}
