% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsclient.R
\name{download_fq_dataset_attachment_rs}
\alias{download_fq_dataset_attachment_rs}
\title{download_fq_dataset_attachment_rs}
\usage{
download_fq_dataset_attachment_rs(
  client,
  attachment_name,
  outpath,
  dataset_id = NULL,
  dataset_name = NULL
)
}
\arguments{
\item{client}{ReadStore client}

\item{attachment_name}{Attachment name to download}

\item{outpath}{Path to write the attachment}

\item{dataset_id}{Project ID to return}

\item{dataset_name}{Project name to return}
}
\description{
Download dataset attachment from the ReadStore API
Write the attachment to a file
ID or Name must be provided
}
