% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstore.R
\name{download_project_attachment}
\alias{download_project_attachment}
\title{download_project_attachment}
\usage{
download_project_attachment(
  client,
  attachment_name,
  project_id = NULL,
  project_name = NULL,
  outpath = NULL
)
}
\arguments{
\item{client}{ReadStore client}

\item{attachment_name}{Attachment name to download}

\item{project_id}{Project ID to return}

\item{project_name}{Project name to return}

\item{outpath}{Path to save the attachment}
}
\description{
Download attachment from a project
ID or Name of project be provided
Also provide the attachment name to download
}
