% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsclient.R
\name{download_project_attachment_rs}
\alias{download_project_attachment_rs}
\title{download_project_attachment_rs}
\usage{
download_project_attachment_rs(
  client,
  attachment_name,
  outpath,
  project_id = NULL,
  project_name = NULL
)
}
\arguments{
\item{client}{ReadStore client}

\item{attachment_name}{Attachment name to download}

\item{outpath}{Path to write the attachment}

\item{project_id}{Project ID to return}

\item{project_name}{Project name to return}
}
\description{
Download project attachment from the ReadStore API
Write the attachment to a file
ID or Name must be provided
}
