% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstore.R
\name{get_client}
\alias{get_client}
\title{get_client}
\usage{
get_client(
  config_dir = "~/.readstore",
  username = NULL,
  token = NULL,
  host = "http://localhost",
  port = 8000,
  return_type = "json",
  fastq_extensions = c(".fastq", ".fastq.gz", ".fq", ".fq.gz")
)
}
\arguments{
\item{config_dir}{Path to the configuration directory to load the configuration from}

\item{username}{Username to use for authentication}

\item{token}{Token to use for authentication}

\item{host}{Host to connect to}

\item{port}{Port to connect to}

\item{return_type}{The return type to use for the API calls}

\item{fastq_extensions}{FASTQ file extensions to use}
}
\value{
ReadStore client (list)
}
\description{
Get ReadStore client to interact with ReadStore API.
Providing username and token will take precedence over configuration file.
Environment variables will take precedence over configuration file.
}
