% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsclient.R
\name{get_fastq_dataset_rs}
\alias{get_fastq_dataset_rs}
\title{get_fastq_dataset_rs}
\usage{
get_fastq_dataset_rs(client, dataset_id = NULL, dataset_name = NULL)
}
\arguments{
\item{client}{ReadStore client}

\item{dataset_id}{Dataset ID to return}

\item{dataset_name}{Dataset name to return}
}
\value{
json object (list) with fastq dataset
}
\description{
Get fastq dataset from the ReadStore API
ID or Name must be provided
}
