% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsclient.R
\name{get_fq_file_rs}
\alias{get_fq_file_rs}
\title{get_fq_file_rs}
\usage{
get_fq_file_rs(client, fq_file_id)
}
\arguments{
\item{client}{ReadStore client}

\item{fq_file_id}{id of the fastq file}
}
\value{
json object (list) with fastq file
}
\description{
Get a fastq file from the ReadStore API
}
