% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsclient.R
\name{get_fq_file_upload_path_rs}
\alias{get_fq_file_upload_path_rs}
\title{get_fq_file_upload_path_rs}
\usage{
get_fq_file_upload_path_rs(client, fq_file_id)
}
\arguments{
\item{client}{ReadStore client}

\item{fq_file_id}{id of the fastq file}
}
\value{
character with the upload path
}
\description{
Get upload path for a fastq file from the ReadStore API
}
