% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstore.R
\name{get_metadata_df}
\alias{get_metadata_df}
\title{get_metadata_df}
\usage{
get_metadata_df(input_list)
}
\arguments{
\item{input_list}{List of metadata entries}
}
\value{
data.frame with metadata entries
}
\description{
Return metadata entries from a list and cast to a data frame
Metadata is key value pairs and key values are used to
construct the data.frame names
}
