% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstore.R
\name{get_pro_data}
\alias{get_pro_data}
\title{get_pro_data}
\usage{
get_pro_data(
  client,
  pro_data_id = NULL,
  name = NULL,
  version = NULL,
  dataset_id = NULL,
  dataset_name = NULL
)
}
\arguments{
\item{client}{ReadStore client}

\item{pro_data_id}{ProData ID to return}

\item{name}{Name of ProData entry to filter}

\item{version}{Version of ProData entry to filter}

\item{dataset_id}{Dataset ID to filter}

\item{dataset_name}{Dataset name to filter}
}
\value{
(json) ProData entry
}
\description{
Get single Processed Data (ProData) entry from the ReadStore API
Return by ProData Id or combination of Name and Dataset ID or Name
If version is NULL, return the latest valid version
}
