% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstore.R
\name{get_project}
\alias{get_project}
\title{get_projects}
\usage{
get_project(client, project_id = NULL, project_name = NULL, return_type = NULL)
}
\arguments{
\item{client}{ReadStore client}

\item{project_id}{Project ID to return}

\item{project_name}{Project name to return}

\item{return_type}{The return type (currently only json)}
}
\value{
json object (list) with project
}
\description{
Get project from the ReadStore API
ID or Name of project must be provided
}
