% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsclient.R
\name{get_project_rs}
\alias{get_project_rs}
\title{get_project_rs}
\usage{
get_project_rs(client, project_id = NULL, project_name = NULL)
}
\arguments{
\item{client}{ReadStore client}

\item{project_id}{Project ID to return}

\item{project_name}{Project name to return}
}
\value{
json object (list) with project
}
\description{
Get project from the ReadStore API
ID or Name must be provided
}
