% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsclient.R
\name{get_rs_client}
\alias{get_rs_client}
\title{get_rs_client}
\usage{
get_rs_client(username, token, endpoint_url)
}
\arguments{
\item{username}{username}

\item{token}{token to authenticate}

\item{endpoint_url}{Endpoint to check}
}
\value{
list with username, token and endpoint_url defining the ReadStore client
}
\description{
Get a rs client object (client template)
}
