% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstore.R
\name{list_datasets}
\alias{list_datasets}
\title{list_datasets}
\usage{
list_datasets(
  client,
  project_id = NULL,
  project_name = NULL,
  return_type = "data.frame"
)
}
\arguments{
\item{client}{ReadStore client}

\item{project_id}{Project ID to filter datasets for}

\item{project_name}{Project name to filter datasets for}

\item{return_type}{The return type (list | data.frame)}
}
\value{
data.frame or list of datasets
}
\description{
Get list of fastq datasets from the ReadStore API
}
