% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstore.R
\name{list_datasets_metadata}
\alias{list_datasets_metadata}
\title{list_datasets_metadata}
\usage{
list_datasets_metadata(
  client,
  project_id = NULL,
  project_name = NULL,
  return_type = "data.frame"
)
}
\arguments{
\item{client}{ReadStore client}

\item{project_id}{Subset metadata for a project ID}

\item{project_name}{Subset metadata for a project name}

\item{return_type}{The return type (data.frame | list)}
}
\value{
data.frame or list of datasets of metadata
}
\description{
Return metadata for datasets from the ReadStore API
Order of returned metadata is the same as the dataset list
}
\details{
Return a list of metadata entries for each dataset

or return a data.frame with metadata entries
Here metadata keys will be cast to column names (wide format)
}
