% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsclient.R
\name{list_fq_files_rs}
\alias{list_fq_files_rs}
\title{list_fq_files_rs}
\usage{
list_fq_files_rs(client)
}
\arguments{
\item{client}{ReadStore client}
}
\value{
(json) list of projects (list objects)
}
\description{
List fastq files from the ReadStore API
}
