% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstore.R
\name{list_pro_data}
\alias{list_pro_data}
\title{list_pro_data}
\usage{
list_pro_data(
  client,
  project_id = NULL,
  project_name = NULL,
  dataset_id = NULL,
  dataset_name = NULL,
  name = NULL,
  data_type = NULL,
  include_archived = FALSE
)
}
\arguments{
\item{client}{ReadStore client}

\item{project_id}{Project ID to filter}

\item{project_name}{Project name to filter}

\item{dataset_id}{Dataset ID to filter}

\item{dataset_name}{Dataset name to filter}

\item{name}{Name of ProData entry to filter}

\item{data_type}{Data type of ProData entry to filter}

\item{include_archived}{Include archived entries (bool)}
}
\value{
(json) list of ProData entries
}
\description{
Get list of ProData entries from the ReadStore API
include_archived flag also shows archived entries
}
