% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstore.R
\name{list_pro_data_metadata}
\alias{list_pro_data_metadata}
\title{list_pro_data_metadata}
\usage{
list_pro_data_metadata(
  client,
  project_id = NULL,
  project_name = NULL,
  dataset_id = NULL,
  dataset_name = NULL,
  name = NULL,
  data_type = NULL,
  include_archived = FALSE,
  return_type = "data.frame"
)
}
\arguments{
\item{client}{ReadStore client}

\item{project_id}{Project ID to filter}

\item{project_name}{Project name to filter}

\item{dataset_id}{Dataset ID to filter}

\item{dataset_name}{Dataset name to filter}

\item{name}{Name of ProData entry to filter}

\item{data_type}{Data type of ProData entry to filter}

\item{include_archived}{Include archived entries (bool)}

\item{return_type}{The return type (list | data.frame)}
}
\value{
data.frame or list of datasets
}
\description{
Return metadata for processed data (pro_data) from the ReadStore API
Order of the returned metadata is the same as the pro_data list
}
\details{
Return a list of pro_data entries for each dataset

or return a data.frame with metadata entries
Here metadata keys will be cast to column names (wide format)
}
