% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstore.R
\name{list_projects}
\alias{list_projects}
\title{list_projects}
\usage{
list_projects(client, return_type = NULL)
}
\arguments{
\item{client}{ReadStore client}

\item{return_type}{The return type (currently only json)}
}
\value{
(json) list of projects (list objects)
}
\description{
Get list of projects from the ReadStore API
}
