% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstore.R
\name{list_projects_metadata}
\alias{list_projects_metadata}
\title{list_projects_metadata}
\usage{
list_projects_metadata(client, return_type = "data.frame")
}
\arguments{
\item{client}{ReadStore client}

\item{return_type}{The return type (data.frame | list)}
}
\value{
data.frame or list of datasets of metadata
}
\description{
Return metadata for projects from the ReadStore API
Order of returned projects is the same as the dataset list
}
\details{
Return a list of metadata entries for each project

or return a data.frame with metadata entries
Here metadata keys will be cast to column names (wide format)
}
