% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsclient.R
\name{list_projects_rs}
\alias{list_projects_rs}
\title{list_projects_rs}
\usage{
list_projects_rs(client)
}
\arguments{
\item{client}{ReadStore client}
}
\value{
(json) list of projects (list objects)
}
\description{
Get list of projects from the ReadStore API
}
