% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstore.R
\name{update_dataset}
\alias{update_dataset}
\title{update_dataset}
\usage{
update_dataset(
  client,
  dataset_id,
  dataset_name = NULL,
  description = NULL,
  project_ids = NULL,
  project_names = NULL,
  metadata = NULL
)
}
\arguments{
\item{client}{ReadStore client}

\item{dataset_id}{Dataset ID}

\item{dataset_name}{Dataset name (NULL or string)}

\item{description}{Dataset description (NULL or string)}

\item{project_ids}{List of project IDs (NULL or vector)}

\item{project_names}{List of project names (NULL or vector)}

\item{metadata}{Named list of metadata (NULL or list)}
}
\value{
Created dataset object
}
\description{
Update a dataset defined by dataset_id
NULL values in the arguments will not be updated
}
