% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstore.R
\name{update_project}
\alias{update_project}
\title{update_project}
\usage{
update_project(
  client,
  project_id,
  project_name = NULL,
  description = NULL,
  metadata = NULL,
  dataset_metadata_keys = NULL
)
}
\arguments{
\item{client}{ReadStore client}

\item{project_id}{Dataset ID}

\item{project_name}{Dataset name (NULL or string)}

\item{description}{Dataset description (NULL or string)}

\item{metadata}{Named list of metadata (NULL or list)}

\item{dataset_metadata_keys}{Named list of metadata (NULL or vector)}
}
\value{
Created dataset object
}
\description{
Update a dataset defined by dataset_id
NULL values in the arguments will not be updated
}
