% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsclient.R
\name{upload_pro_data_rs}
\alias{upload_pro_data_rs}
\title{upload_pro_data_rs}
\usage{
upload_pro_data_rs(
  client,
  name,
  pro_data_path,
  data_type,
  metadata = list(),
  description = "",
  dataset_id = NULL,
  dataset_name = NULL
)
}
\arguments{
\item{client}{ReadStore client}

\item{name}{Name of ProData entry}

\item{pro_data_path}{Path to the ProData file}

\item{data_type}{Data type of the ProData entry}

\item{metadata}{Metadata key value list}

\item{description}{Description of the ProData entry}

\item{dataset_id}{Dataset ID to attach the ProData entry}

\item{dataset_name}{Dataset name to attach the ProData entry}
}
\description{
Upload ProData entry using the ReadStore API
Dataset ID or Name must be provided to select the dataset to attach
}
\details{
Return 403 if the user does not have permission to upload ProData
}
