% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsclient.R
\name{validate_charset}
\alias{validate_charset}
\title{validate_charset}
\usage{
validate_charset(query_str)
}
\arguments{
\item{query_str}{String to validate}
}
\value{
TURE if valid charset, FALSE otherwise
}
\description{
Check if the query string contains only allowed characters
Allowed characters are: 0-9, a-z, A-Z, _, -, ., @
}
