% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsclient.R
\name{validate_metadata}
\alias{validate_metadata}
\title{validate_metadata}
\usage{
validate_metadata(metadata)
}
\arguments{
\item{metadata}{List if metadata entries}
}
\description{
Check if metadata keys are valid
Allowed characters are: 0-9, a-z, A-Z, _, -, ., @
Must not be empty
Must not be a reserved keyword
}
