% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsconfig.R
\name{write_rs_config}
\alias{write_rs_config}
\title{write_rs_config}
\usage{
write_rs_config(
  filename,
  username,
  token,
  endpoint_url,
  fastq_extensions,
  output
)
}
\arguments{
\item{filename}{Path to configuration file}

\item{username}{Username}

\item{token}{Token}

\item{endpoint_url}{Endpoint URL}

\item{fastq_extensions}{Fastq extensions}

\item{output}{Output format}
}
\description{
Write ReadStore configuration to file in INI format
Set permissions to 0600 (read/write only by owner)
}
