% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstore.R
\name{create_dataset}
\alias{create_dataset}
\title{create_dataset}
\usage{
create_dataset(
  client,
  name,
  description = "",
  project_ids = c(),
  project_names = c(),
  metadata = list()
)
}
\arguments{
\item{client}{ReadStore client}

\item{name}{Dataset name}

\item{description}{Dataset description}

\item{project_ids}{Vector of project IDs}

\item{project_names}{Vector of project names}

\item{metadata}{Named list of metadata}

\item{qc_passed}{QC Pass}

\item{paired_end}{Paired End}

\item{index_read}{Index read}
}
\value{
Created dataset object
}
\description{
Create a new dataset by calling create_fastq_dataset_rs
}
