% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstore.R
\name{list_projects}
\alias{list_projects}
\title{list_projects}
\usage{
list_projects(client, return_type = "data.frame")
}
\arguments{
\item{client}{ReadStore client}

\item{return_type}{The return type  (list | data.frame)}
}
\value{
data.frame or list of projects
}
\description{
Get list of projects from the ReadStore API
}
