% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstore.R
\name{upload_fastq}
\alias{upload_fastq}
\title{upload_fastq}
\usage{
upload_fastq(client, fastq)
}
\arguments{
\item{client}{ReadStore client}

\item{fastq}{FASTQ files to upload as a vector}
}
\description{
Upload FASTQ files to ReadStore
}
