% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsclient.R
\name{upload_fastq_rs}
\alias{upload_fastq_rs}
\title{upload_fastq_rs}
\usage{
upload_fastq_rs(client, fastq_files)
}
\arguments{
\item{client}{ReadStore client}

\item{fastq_files}{vector of fastq files to upload}
}
\description{
Upload fastq files to the ReadStore API
}
